from dev/src * Houdini Engine/HAPI: changed all of the time related variables to doubles to match time being specified as an fpreal within Houdini (apart from float builds). This fixes issues with lack of precision showing up in frame numbers (instead of 101, getting frame 100.999999), which was causing issues for customers. * Houdini Engine for 3ds Max: updated the HAPI bindings * Houdini Engine for 3ds Max: swept the plugin to change time variables from float to double in order to reflect the change in HAPI * Houdini Engine for Unreal: updated HAPI bindings * HoudiniEngineSample: updated HAPI bindings * Houdini Engine for Maya: updated HAPI bindings * Houdini Engine for Maya: updated the plugin to use double to represent time instead of float, in order to reflect the changes in HAPI * Houdini Engine for Unity: updated the GetTime and SetTime methods to use doubles instead of floats to reflect the change in HAPI

+++ Asset.C
@@ -922,7 +922,7 @@ Asset::setTime(const MTime &mayaTime)
     // Houdini's "frame 1" is "0 seconds", but Maya's "frame 0" is "0 seconds".
     // So we need to offset the time by 1.
     MTime hapiTime        = myTime - MTime(1, MTime::uiUnit());
-    float hapiTimeSeconds = (float)hapiTime.as(MTime::kSeconds);
+    double hapiTimeSeconds = hapiTime.as(MTime::kSeconds);
     HoudiniApi::SetTime(Util::theHAPISession.get(), hapiTimeSeconds);
 }
